noticket: Minor `YqlStatementPart` cleanup
- Stricter default implementation of `YqlStatementPart.combine()` that throws `UnsupportedOperationException`
- Removed the (incorrect and strange) `YqlOrderBy.combine()` implementation that ignores `this.sortKeys`
*and* does not attempt to de-duplicate sort keys, allowing multiple occurrences of the same sort key, even
with a different sort direction.
Known YOJ users are unlikely to have used the `YqlOrderBy.combine()` functionality, ever.
- Better function signature for `YqlStatement.mergeParts()` that combines `YqlStatementPart`s of the same type.
It no longer takes a `Stream` of `YqlStatementPart`s but a `Collection` of them, which is appropriate for
most usages. Implementations of custom statements (especially inheriting from `PredicateStatement`) will need
to eventually transition to the new `mergeParts` variant.
- Improved consistency in `YqlLimit` and `YqlView`: use `getXyz()` POJO-style getters instead of `xyz()`,
deprecate the confusingly named `YqlLimit.size()` method in favor of explicit `getLimit()` and `getOffset()`
getters which are more general.