noticket: FIX: Unexpected NPE when trying to validate conversion result in `CustomValueTypes.preconvert()`
This happens when a custom-value converter erroneously returns `null` from `ValueConverter.toColumn()`,
e.g. because of a bad `null`-returning `toString()` implementation in a string-value type.

This should not happen in `CustomValueTypes.postconvert()` because the `null -> null` logic is applied
before that, and a null `value` should not get into `CustomValueTypes.postconvert()` **at all**.