#77: Remove deprecation from user-visible parts of the Projections API (for now) + Improve messages for dry-run mode transactions
We remove `@Deprecated` annotation and deprecation warning logs from user-visible parts
of the Projections API: `t.y.y.r.d.{Entity.createProjections, EntityList}`, `t.y.y.r.d.projection.*` 
(except for `@InternalApi`s). 

We add concrete docs on which parts of the APIs will be moved to a different library and/or
package, and which internal Projection-related APIs will be removed entirely (or at least *may* be.)

Drive-by change: Improve transaction log message and ordinary log message when we roll back
a dry-run transaction. It's not caused by any "inconsistency", and is in fact entirely normal and indended.

Co-authored-by: Alexander Lavrukov <lavrukov@yandex-team.ru>
Co-authored-by: Nikolai Amelichev <nickamelichev@gmail.com>